crawler/app/requirements.txt
APScheduler
pillow
requests
pyyaml